Restuructured bc_protocol to increase modulaitry
