+++ crates/core/src/bc_protocol/logout.rs
@@ -1,0 +1,12 @@
+use super::{BcCamera, Result};
+
+impl BcCamera {
+    /// Logout from the camera
+    pub fn logout(&mut self) -> Result<()> {
+        if self.logged_in {
+            // TODO: Send message ID 2
+        }
+        self.logged_in = false;
+        Ok(())
+    }
+}
